Move project specific config into .env file for Makefile deploys
Remove name, project, and region from the Makefile and store them in a
gitignored `.env` file instead. They are not really secret but that
makes the file general enough that other people could fork the repos and
use it unchanged by just adding their own config to a `.env` file.